+++ src/pages-tab/components/page-fields/page-fields.less
@@ -3,4 +3,7 @@
 page-fields {
   display: block;
   /* padding: @grid-gutter-width; */
+  .type-error {
+    margin-bottom: 10px;
+  }
 }
